0.4

+++ handlers/downloader.py
@@ -1,5 +1,6 @@
 import yt_dlp
 import os
+
 from aiogram import Router, F, types
 from aiogram.types import Message, CallbackQuery
 from aiogram.types import FSInputFile
@@ -9,33 +10,43 @@ router = Router()
 video_dir = 'for_files'
 
 options = {
-    'format': 'ba',
+    'format': 'mp3/bestaudio/best',
     'outtmpl': f'{video_dir}/%(title)s.mp3',
-    'playlist_items': "0:1"
+    'noplaylist': True,
 }
 
 
 @router.message(F.text, flags={"long_operation": "upload_video_note"})
 async def downloader(message: Message):
-    url = message.text
-    tiny_url = url.split('&')
-
-    with yt_dlp.YoutubeDL(options) as ydl:
+    ydl = yt_dlp.YoutubeDL(options)
+    if message.entities is not None:
+        url = message.text
         info = ydl.extract_info(url, download=False)
 
-    if len(tiny_url) > 1:
-        video_title = info['entries'][0]['title']
     else:
-        video_title = info.get('title')
+        info = ydl.extract_info(f"ytsearch1:{message.text}", download=False)['entries'][0]
+        url = info.get('url')
+        ydl.download(f"ytsearch1:{message.text}")
 
-    video_file = f'{video_title}.mp3'
-    video_path = f'{video_dir}\{video_file}'
+    video_title = info.get('title')
+    video_duration = info.get('duration')
+    video_file = f"{video_title}.mp3"
+    video_path = f"{video_dir}\\{video_file}"
 
-    with yt_dlp.YoutubeDL(options) as ydl:
+    if video_duration > 1200:
+        await message.answer(
+            text="Слишком долгое видео, пожалейте бота :["
+        )
+
+    else:
         ydl.download(url)
 
-    audio_final = FSInputFile(video_path)
+        final_audio = FSInputFile(video_path)
 
-    await message.reply_audio(audio_final)
+        await message.answer_audio(
+                audio=final_audio,
+                duration=video_duration,
+                title=video_title,
+        )
 
-    os.remove(video_path)
+        os.remove(video_path)
